+++ resources/views/jquery/items.blade.php
@@ -5,11 +5,11 @@
 <!-- head of items -->
 @if(count($items))
 <strong>
-    @if($state == \App\Model\Enum\ItemState::UNCHECKED)
+    @if($state == \App\Model\Enum\ItemState::UNCHECKED && count($items))
     <button
             title="{{ $trans->get('ShowItems.CheckAll.Title') }}"
             class="move_item"
-            onclick="updateAllItems()">
+            onclick="updateAllItems('{{ route('put.items.check', ['id' => $items[0]['wallet']]) }}', '{{ route('get.items.wallet', ['id' => $items[0]['wallet']]) }}', '{{ App\Model\Enum\ItemState::UNCHECKED }}', '{{$trans->get('PrintItems.MoveAllToArchive', 'Do you really want to move this items to archive?')}}')">
         <b>&#10004;</b>
     </button>
     @endif
@@ -34,19 +34,12 @@
     <div id="itemDiv_{{ $item['id'] }}" class="item{{ (!$item['active'] ? ' old' : $item['odepsat'] ? ' odepsat' : '') }}">
         <div class="buttons">
             <button title="{{ $trans->get('PrintItems.DeleteItemTitle', 'Delete') }}" class="delete_item"
-                             onclick="deleteItem('{{ $item['id'] }}', '{{ route('delete.item', ['id' => $item['id']]) }}')"><b>&times;</b></button>
+                             onclick="deleteItem('{{ $item['id'] }}', '{{ route('delete.item', ['id' => $item['id']]) }}', '{{$trans->get('PrintItems.DeleteItem.Confirm', 'Do you really want to delete this item?')}}')"><b>&times;</b></button>
             @if($state == \App\Model\Enum\ItemState::UNCHECKED)
             <button title="{{ $trans->get('PrintItems.CheckedItemTitle', 'Move to archive') }}" class="move_item"
                              onclick="updateItemRead('{{ $item['id'] }}', '{{ route('put.item.check', ['id' => $item['id']]) }}')"><b>&#10004;</b></button>
             <button title="{{ $trans->get('PrintItems.UpdateItemTitle', 'Update') }}" class="updateItem"
                              onclick="updateItemMakeForm('{{ $item['id'] }}', '{{ route('get.item.HTML', ['id' => $item['id']]) }}')"></button>
-            <!--
-            <form method="POST" action="' . route('get.item.update', ['id' => $item['id']]) . '">
-                {{ csrf_field() }}
-                <input type="hidden" name="_method" value="get">
-                <button title="{{ $trans->get('PrintItems.UpdateItemTitle') }}'" class="updateItem" onclick="updateItemMakeForm({{ $item['id'] }})"></button>
-            </form>
-            -->
             @endif
             </div>
